feat: 디테일 페이지 params:id 기능 구현
