+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { Routes, Route } from "react-router-dom";
 import { Home, Detail } from "./pages";
-import { DataDummy, Layout, NotFound } from "./components";
+import { DataDummy, Layout, NotFound, TabContent } from "./components";
 
 export default function App(): JSX.Element {
   return (
@@ -12,10 +12,11 @@ export default function App(): JSX.Element {
             <Layout />
             <DataDummy />
           </>
-        }>
+        }
+      >
         <Route index element={<Home />} />
-        <Route path="/detail" element={<Detail />}>
-          {/* <Route path="/:id" element={<></>} /> */}
+        <Route path="detail" element={<Detail />}>
+          <Route path=":id" element={<TabContent />} />
         </Route>
         <Route path="*" element={<NotFound />} />
       </Route>
